changes for the moved generated App service
