+++ src/Graviton/CoreBundle/Document/Dummy.php
@@ -1,0 +1,44 @@
+<?php
+/**
+ * dummy document
+ */
+
+namespace Graviton\CoreBundle\Document;
+
+/**
+ * Version
+ *
+ * @author   List of contributors <https://github.com/libgraviton/graviton/graphs/contributors>
+ * @license  https://opensource.org/licenses/MIT MIT License
+ * @link     http://swisscom.ch
+ */
+class Dummy
+{
+
+    /**
+     * @var string
+     */
+    private $id;
+
+    /**
+     * get Id
+     *
+     * @return mixed Id
+     */
+    public function getId()
+    {
+        return $this->id;
+    }
+
+    /**
+     * set Id
+     *
+     * @param mixed $id id
+     *
+     * @return void
+     */
+    public function setId($id)
+    {
+        $this->id = $id;
+    }
+}
