+++ src/Graviton/CoreBundle/Model/Version.php
@@ -8,8 +8,6 @@ namespace Graviton\CoreBundle\Model;
 use Graviton\RestBundle\Model\DocumentModel;
 
 /**
- * Model based on Graviton\RestBundle\Model\DocumentModel.
- *
  * @author   List of contributors <https://github.com/libgraviton/graviton/graphs/contributors>
  * @license  https://opensource.org/licenses/MIT MIT License
  * @link     http://swisscom.ch
