+++ src/Graviton/CoreBundle/Resources/config/doctrine/Dummy.mongodb.yml
@@ -1,0 +1,6 @@
+Graviton\CoreBundle\Document\Dummy:
+  fields:
+    id:
+      type: string
+      id: true
+      strategy: UUID
