+++ src/Graviton/SecurityBundle/Resources/config/services.yml
@@ -64,11 +64,11 @@ services:
 
     graviton.security.authentication.provider.model.noop:
         class: Graviton\SecurityBundle\User\Model\NullModel
-#        calls:
-#          -
-#            method: setRepository
-#            arguments:
-#              - "@gravitondyn.app.repository.app"
+        calls:
+          -
+            method: setRepository
+            arguments:
+              - "@graviton.core.repository.dummy"
 
     #
     #    Commands
@@ -96,13 +96,13 @@ services:
     graviton.security.model.whoami:
         class: Graviton\SecurityBundle\Model\WhoAmI
         parent: graviton.rest.model
-#        arguments:
-#          - "@graviton.rql.visitor.mongodb"
-#        calls:
-#          -
-#            method: setRepository
-#            arguments:
-#              - "@gravitondyn.app.repository.app"
+        arguments:
+          - "@graviton.rql.visitor.mongodb"
+        calls:
+          -
+            method: setRepository
+            arguments:
+              - "@graviton.core.repository.dummy"
 
     graviton.security.controller.whoami:
         class: Graviton\SecurityBundle\Controller\WhoAmIController
